Add authentication against another database
mongotail only supports authentication against the database specified in
the db address. But MongoDB allows too to authenticate against a
different database, e.g. via the "--authenticationDatabase" option in
the mongo binary.

This commit attempts to add support in mongotail to authenticate against
a different database than the specified in the db address. A new
optional argument "-b/--authenticationDatabase" has been added to
specify that database. If not provided and authentication is required,
mongotail will use the database specified in the db address.